p-v部分接口提取
