+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".MainFrame.MainActivity"
             android:label="@string/app_name"
             android:screenOrientation="fullSensor"
             android:theme="@style/DrawerLayoutTheme">
@@ -31,7 +31,7 @@
             android:name=".PlayerModule.PlayerActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenSize|screenLayout|smallestScreenSize|uiMode"
             android:launchMode="singleTop"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName=".MainFrame.MainActivity"
             android:screenOrientation="sensorLandscape"
             android:theme="@style/MyTheme.FullScreen">
             <intent-filter>
@@ -70,7 +70,7 @@
             android:exported="false">
         </service>
         <!--最先启动的activity，用于解决烦人的黑白屏问题-->
-        <activity android:name=".WelcomeActivity"
+        <activity android:name=".MainFrame.WelcomeActivity"
             android:theme="@style/StartActivityTransparent">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
