+++ app/src/main/java/com/studyjams/mdvideo/HistoryVideo/VideoPlayHistoryCursorAdapter.java
@@ -1,4 +1,4 @@
-package com.studyjams.mdvideo.Adapter;
+package com.studyjams.mdvideo.HistoryVideo;
 
 import android.content.Context;
 import android.database.Cursor;
@@ -23,7 +23,7 @@ import java.util.TimeZone;
 /**
  * Created by syamiadmin on 2016/7/12.
  */
-public class VideoLocalCursorAdapter extends RecyclerViewCursorAdapter<VideoLocalCursorAdapter.VideoViewHolder> {
+public class VideoPlayHistoryCursorAdapter extends RecyclerViewCursorAdapter<VideoPlayHistoryCursorAdapter.VideoViewHolder> {
 
     private static final String TAG = "LocalVideoCursorAdapter";
     private List<Video> mVideoData;
@@ -32,20 +32,15 @@ public class VideoLocalCursorAdapter extends RecyclerViewCursorAdapter<VideoLoca
      * Constructor.
      * @param context The Context the Adapter is displayed in.
      */
-    public VideoLocalCursorAdapter(Context context) {
+    public VideoPlayHistoryCursorAdapter(Context context) {
         super(context);
 
-        setupCursorAdapter(null, 0, R.layout.video_local_list_item, false);
+        setupCursorAdapter(null, 0, R.layout.video_play_history_item, false);
         mVideoData = new ArrayList<>();
 
         /**Format time**/
         mDateFormat = new SimpleDateFormat("HH:mm:ss", Locale.getDefault());
         mDateFormat.setTimeZone(TimeZone.getTimeZone("GMT+0:00"));
-    }
-
-    @Override
-    public int getItemViewType(int position) {
-        return super.getItemViewType(position);
     }
 
     /**
@@ -54,6 +49,7 @@ public class VideoLocalCursorAdapter extends RecyclerViewCursorAdapter<VideoLoca
      * @return
      */
     public Video getItemData(int position){
+
         return mVideoData.get(position);
     }
 
@@ -90,27 +86,36 @@ public class VideoLocalCursorAdapter extends RecyclerViewCursorAdapter<VideoLoca
         public final TextView mTitle;
         public final TextView mInfo;
         public final TextView mSize;
+        public final TextView mDate;
 
         public VideoViewHolder(View view) {
             super(view);
 
-            mThumbnail = (ImageView) view.findViewById(R.id.local_list_item_image);
-            mTitle = (TextView) view.findViewById(R.id.local_list_item_title);
-            mInfo = (TextView) view.findViewById(R.id.local_list_item_info);
-            mSize = (TextView) view.findViewById(R.id.local_list_item_size);
+            mThumbnail = (ImageView) view.findViewById(R.id.play_history_item_image);
+            mTitle = (TextView) view.findViewById(R.id.play_history_item_title);
+            mInfo = (TextView) view.findViewById(R.id.play_history_item_info);
+            mSize = (TextView) view.findViewById(R.id.play_history_item_size);
+            mDate = (TextView) view.findViewById(R.id.play_history_item_time);
         }
 
         @Override
         public void bindCursor(Cursor cursor) {
 
-            final Video video = Video.from(cursor);
+            Video video = Video.from(cursor);
 
             /**save data for click event**/
             mVideoData.add(getAdapterPosition(),video);
+            String str;
+            if(video.isPlayCompleted()){
+                str = mContext.getResources().getString(R.string.player_play_end);
+            }else{
+                str = mContext.getResources().getString(R.string.player_play_history) + mDateFormat.format(video.getPlayDuration());
+            }
 
             mTitle.setText(video.getTitle());
-            mInfo.setText(mDateFormat.format(video.getDuration()));
+            mInfo.setText(str);
             mSize.setText(Formatter.formatFileSize(mContext,video.getSize()));
+            mDate.setText(video.getCreatedDate());
             Tools.LoadNormalImage(mContext,video.getPath(),mThumbnail);
         }
     }
