+++ app/src/main/java/com/studyjams/mdvideo/MainFrame/WelcomeActivity.java
@@ -1,4 +1,4 @@
-package com.studyjams.mdvideo;
+package com.studyjams.mdvideo.MainFrame;
 
 import android.app.Activity;
 import android.content.Intent;
