+++ app/src/main/res/layout/app_bar_main.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.studyjams.mdvideo.MainActivity">
+    tools:context=".MainFrame.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
