+++ app/src/main/res/values/strings.xml
@@ -7,6 +7,9 @@
     <string name="navigation_drawer_close">Close navigation drawer</string>
 
     <string name="action_settings" translatable="false">设置</string>
+    <!--local-->
+    <string name="local_no_video">没有可播放的视频~</string>
+
 
     <!--导航菜单-->
     <string name="menu_video_local">本地视频</string>
