+++ app/src/main/res/values/styles.xml
@@ -5,7 +5,6 @@
         <item name="colorPrimary">@color/primary</item>
         <item name="colorPrimaryDark">@color/primaryDark</item>
         <item name="colorAccent">@color/accent</item>
-        <item name="android:windowContentOverlay">@null</item>
     </style>
 
     <!--处理黑白屏问题-->
